--- src/targets/gpu/include/migraph/gpu/lrn.hpp
@@ -1,39 +1,0 @@
-#ifndef MIGRAPH_GUARD_RTGLIB_LRN_HPP
-#define MIGRAPH_GUARD_RTGLIB_LRN_HPP
-
-#include <migraph/gpu/lowering.hpp>
-#include <migraph/manage_ptr.hpp>
-#include <migraph/instruction.hpp>
-#include <migraph/operators.hpp>
-#include <migraph/generate.hpp>
-#include <migraph/shape_for_each.hpp>
-#include <migraph/config.hpp>
-#include <migraph/gpu/miopen.hpp>
-#include <migraph/gpu/hip.hpp>
-#include <migraph/dfor.hpp>
-#include <migraph/gpu/device/contiguous.hpp>
-#include <migraph/gpu/device/add.hpp>
-#include <migraph/iterator_for.hpp>
-#include <migraph/gpu/rocblas.hpp>
-#include <migraph/gpu/context.hpp>
-#include <utility>
-
-namespace migraph {
-inline namespace MIGRAPH_INLINE_NS {
-namespace gpu {
-
-struct miopen_lrn
-{
-    shared<lrn_descriptor> ldesc;
-    std::string name() const { return "gpu::lrn"; }
-    shape compute_shape(const std::vector<shape>& inputs) const;
-    argument
-    compute(context& ctx, const shape& output_shape, const std::vector<argument>& args) const;
-    int output_alias(const std::vector<shape>& shapes) const { return shapes.size() - 1; }
-};
-
-} // namespace gpu
-} // namespace MIGRAPH_INLINE_NS
-} // namespace migraph
-
-#endif
